feat(client): show query results on suggestion screen
