+++ packages/client/screens/SuggestTrackModal.tsx
@@ -1,32 +1,36 @@
 import { useActor, useMachine } from '@xstate/react';
-import { Text } from 'dripsy';
+import { ActivityIndicator, View } from 'dripsy';
+import { FlatList } from 'react-native';
 import React, { useState } from 'react';
 import { ActorRef } from 'xstate';
+import TrackListItem from '../components/Track/TrackListItem';
+import { useSuggestTracks } from '../hooks/musicPlayerHooks';
 import { AppScreenWithSearchBar } from '../components/kit';
 import {
     AppScreenHeaderWithSearchBarMachineEvent,
     AppScreenHeaderWithSearchBarMachineState,
 } from '../machines/appScreenHeaderWithSearchBarMachine';
 import { searchTrackMachine } from '../machines/searchTrackMachine';
+import { assertEventType } from '../machines/utils';
 import { SuggestTrackModalProps } from '../types';
 
 const SuggestTrackModal: React.FC<SuggestTrackModalProps> = ({
     navigation,
 }) => {
+    const { suggestTracks, showActivityIndicatorOnSuggestionsResultsScreen } =
+        useSuggestTracks(exitModal);
     const [screenOffsetY, setScreenOffsetY] = useState(0);
-    const [state] = useMachine(searchTrackMachine, {
+    const [state, sendToSearchTracks] = useMachine(searchTrackMachine, {
         actions: {
-            navigateToResultsPage: ({ tracks }) => {
-                if (tracks === undefined) {
-                    return;
-                }
+            handleTrackPressed: (_, event) => {
+                assertEventType(event, 'PRESS_TRACK');
+                const { trackID } = event;
 
-                navigation.navigate('SuggestTrackResultsModal', {
-                    tracks,
-                });
+                suggestTracks([trackID]);
             },
         },
     });
+    const tracksResults = state.context.tracks;
     const searchBarActor: ActorRef<
         AppScreenHeaderWithSearchBarMachineEvent,
         AppScreenHeaderWithSearchBarMachineState
@@ -34,8 +38,22 @@ const SuggestTrackModal: React.FC<SuggestTrackModalProps> = ({
     const [searchState, sendToSearch] = useActor(searchBarActor);
     const showHeader = searchState.hasTag('showHeaderTitle');
 
+    function exitModal() {
+        navigation.popToTop();
+        navigation.goBack();
+    }
+
     function handleGoBack() {
         navigation.goBack();
+    }
+
+    function handleTrackPress(trackID: string) {
+        return () => {
+            sendToSearchTracks({
+                type: 'PRESS_TRACK',
+                trackID,
+            });
+        };
     }
 
     return (
@@ -50,7 +68,42 @@ const SuggestTrackModal: React.FC<SuggestTrackModalProps> = ({
             sendToSearch={sendToSearch}
             goBack={handleGoBack}
         >
-            <Text sx={{ color: 'white' }}>This is the modal</Text>
+            <FlatList
+                data={tracksResults ?? []}
+                renderItem={({ item: { id, title, artistName }, index }) => (
+                    <View
+                        sx={{
+                            marginBottom: 'm',
+                        }}
+                    >
+                        <TrackListItem
+                            index={index + 1}
+                            title={title}
+                            trackID={id}
+                            artistName={artistName}
+                            onPress={handleTrackPress(id)}
+                        />
+                    </View>
+                )}
+                keyExtractor={(_, index) => String(index)}
+            />
+
+            {showActivityIndicatorOnSuggestionsResultsScreen === true && (
+                <View
+                    sx={{
+                        position: 'absolute',
+                        top: 0,
+                        left: 0,
+                        right: 0,
+                        bottom: 0,
+
+                        justifyContent: 'center',
+                        alignItems: 'center',
+                    }}
+                >
+                    <ActivityIndicator size="large" />
+                </View>
+            )}
         </AppScreenWithSearchBar>
     );
 };
